Implemented parts of fire sim and made steps towards fire propagation

+++ Ported/BucketBrigade/Assets/Component/FireCell.cs
@@ -3,5 +3,5 @@
 [GenerateAuthoringComponent]
 public struct FireCell : IComponentData
 {
-    public float Temperature;
+    public byte Temperature;
 }
